Created Basic Front Page
The home page of the website now has a menu for a user to select
single or multiple, centralized or distributed or both, and simply supported or cantilever.
Then, they can click Start to be taken to the appropriate page.

Home is now added to the top navigation bar, so users can easily return to this page.

BUILD_PAGE.md (an instructions document on how to create new pages) is updated
to reflect the new steps needed to make a page reachable via the home page buttons.