Added Input Validation to singleLoadApp.
Information entered into the initial form for singleLoadApp will now be
checked to ensure all entries are numbers which are in the correct ranges.

If any information is invalid, a descriptive error message will appear
in bold above the submit button.

The submit button will not function unless all entries are valid.

Corrected typos: loacation -> location.