Edit and Delete buttons now disabled if 0 loads.
Edit and Delete buttons will now be disabled when there are no loads left, in multiple load apps.

Also commented out a line in movement bending diagram function for cantileverBeamApp
because it would crash the app when there are no loads
and that line only declared an unused variable so it was most likely safe to remove.