Distributed load position is now the center
In distributedLoadApp and combinedLoadApp, all text displays and inputs now consider the load position to be the center of the load.
Internally, the load location is still stored and treated as the left edge of the load.
It is converted to the center for displays and initial textbox values, then converted back to the left upon the user submitting values.

Also, the initial value for load location when creating a new load in multiple load apps is now the middle of the beam instead of 10.