Added cantilever support type to combinedLoadApp
Users see radio button options on the first page to select simply supported beam or cantilever.
Shear force diagram, bending moment diagram, and support reactions are now updated to work with cantilever.
However, deflection is still incomplete and was not added in this git commit.

Mass of load input textbox now says "Mass per Meter" if the current load is distributed.
This will inform users that the mass textbox controls the mass per length, instead of the total mass of the distributed load.

Support reactions are now displayed on the deflection plot instead of on their own plot.