Added Triangular Load displays + data, but no math
Triangular load type was added.

Users can select the name, location, mass, and length, just like distributed loads.
They may also select whether the higher end of the load is on the left or right side.

At the moment, triangular loads are treated as distributed loads by the formulas.

All triangular loads display as the same height, regardless of mass.
It is assumed that the mass is greater because of density instead of height.

Also separated the add load and edit load forms into a separate component file AddEditForm to clean up the code.
Add and edit methods have been merged using a mode parameter to make it easier to pass into the component.

Fixed a bug where the program was not properly identifying when users change the beam length to put distributed loads out of bounds.

Rearranged some labels so type comes first, then location, then mass.