Improvements
Fixed rerendering issue, so users can now hold left/right arrow key and the load will smoothly move instead of freezing for several seconds.
Scale method has been slightly changed so that the plots move smoothly when the load is smoothly moving.

Made the right-side plots scroll in a column independent of the left side, so users can view the bottom-right plot side-by-side with the load positions if they want to.

Support positions are now only checked for simply supported beam.
If support positions are 50 and the user sets length to 40 while on the cantilever setting, the program does not give an error and
force them to switch back to simply supported in order to enable the support position textboxes so they can resolve the problem.

Fixed floating point errors in load position.

Arrow keys and buttons now move the load by 1% of the beam length instead of by 1 unit.