Added display for exact Y, and coordinates of max.
Users can now hover the mouse over plots to show a grey line indicating the x value they are hovering on,
and (x,y) coordinates of the plot point at that x.

Users can also choose an x by typing in an input field.

Also, the global extreme value is displayed for bending moment and deflection.

Mass can no longer be negative because it complicates the max value finding process
by making it possible for multiple local extrema to be in the graph.