Fixed plots labels on different-sized screens
Plot labels now adjust appropriately when the zoom level is changed or the window is resized.

If supports are exactly on top of each other, their sum is now displayed as a single label
(because separately they are infinite but their sum is finite).

Gravity can now be negative, since negative mass and gravity work properly.
Default gravity is now 1 instead of 9.8, as suggested by a user.