Dynamic deflection plot improved
Input lag is reduced, the scale matches the static deflection plot
but gives 3 times as much space to fit the swaying onscreen. It still doesn't always succeed.

There is still a problem where after moving, the dynamic plot suddenly jumps to a different height, but that seems to be a backend problem.