Changes to mobile Android_Version
No longer need to press start button to start the simulation.

Fixed rotation-detection using wrong axis.

Adjusted dimensions of the display, because length 100 is sent to the backend simulator, not length 9.

Also adjusted the height to comfortably fit a single load of this weight on the screen.

Attempted to add motion control to desktop version on android.